Move typings_v1.d.ts to /nlu
